+++ src/Appwrite/Platform/Tasks/Maintenance.php
@@ -1,20 +1,35 @@
 <?php
 
-global $cli;
+namespace Appwrite\Platform\Tasks;
 
 use Appwrite\Auth\Auth;
 use Appwrite\Event\Certificate;
 use Appwrite\Event\Delete;
 use Utopia\App;
 use Utopia\CLI\Console;
+use Utopia\Database\Database;
 use Utopia\Database\Document;
 use Utopia\Database\DateTime;
 use Utopia\Database\Query;
+use Utopia\Platform\Action;
 
-$cli
-    ->task('maintenance')
-    ->desc('Schedules maintenance tasks and publishes them to resque')
-    ->action(function () {
+class Maintenance extends Action
+{
+    public static function getName(): string
+    {
+        return 'maintenance';
+    }
+
+    public function __construct()
+    {
+        $this
+            ->desc('Schedules maintenance tasks and publishes them to resque')
+            ->inject('dbForConsole')
+            ->callback(fn (Database $dbForConsole) => $this->action($dbForConsole));
+    }
+
+    public function action(Database $dbForConsole): void
+    {
         Console::title('Maintenance V1');
         Console::success(APP_NAME . ' maintenance process v1 has started');
 
@@ -102,6 +117,15 @@ $cli
                 ->trigger();
         }
 
+        function notifyDeleteSchedules($interval)
+        {
+
+            (new Delete())
+                ->setType(DELETE_TYPE_SCHEDULES)
+                ->setDatetime(DateTime::addSeconds(new \DateTime(), -1 * $interval))
+                ->trigger();
+        }
+
         // # of days in seconds (1 day = 86400s)
         $interval = (int) App::getEnv('_APP_MAINTENANCE_INTERVAL', '86400');
         $executionLogsRetention = (int) App::getEnv('_APP_MAINTENANCE_RETENTION_EXECUTION', '1209600');
@@ -110,10 +134,9 @@ $cli
         $usageStatsRetentionHourly = (int) App::getEnv('_APP_MAINTENANCE_RETENTION_USAGE_HOURLY', '8640000'); //100 days
 
         $cacheRetention = (int) App::getEnv('_APP_MAINTENANCE_RETENTION_CACHE', '2592000'); // 30 days
+        $schedulesDeletionRetention = (int) App::getEnv('_APP_MAINTENANCE_RETENTION_SCHEDULES', '86400'); // 1 Day
 
-        Console::loop(function () use ($interval, $executionLogsRetention, $abuseLogsRetention, $auditLogRetention, $usageStatsRetentionHourly, $cacheRetention) {
-            $database = getConsoleDB();
-
+        Console::loop(function () use ($interval, $executionLogsRetention, $abuseLogsRetention, $auditLogRetention, $cacheRetention, $schedulesDeletionRetention, $usageStatsRetentionHourly, $dbForConsole) {
             $time = DateTime::now();
 
             Console::info("[{$time}] Notifying workers with maintenance tasks every {$interval} seconds");
@@ -123,7 +146,9 @@ $cli
             notifyDeleteUsageStats($usageStatsRetentionHourly);
             notifyDeleteConnections();
             notifyDeleteExpiredSessions();
-            renewCertificates($database);
+            renewCertificates($dbForConsole);
             notifyDeleteCache($cacheRetention);
+            notifyDeleteSchedules($schedulesDeletionRetention);
         }, $interval);
-    });
+    }
+}